Username Var and Ownership by File in Roles
This adds the user_name as a var in site.yml, and then uses that var to
specify paths for directories and to specify owners and groups for files
that are created in the roles for vim and git.

This also changes the username in site.yml. I'll probably change that to
be populated with a fact at some point soon.